+++ Code/Mantid/Framework/Algorithms/src/AlignDetectors.cpp
@@ -222,7 +222,7 @@ void AlignDetectors::exec()
 
   // Loop over the histograms (detector spectra)
   PARALLEL_FOR2(inputWS,outputWS)
-  for (int64_t i = 0; i < numberOfSpectra; ++i)
+  for (int64_t i = 0; i < int64_t(numberOfSpectra); ++i)
   {
     PARALLEL_START_INTERUPT_REGION
     try {
@@ -300,7 +300,7 @@ void AlignDetectors::execEvent()
   Progress progress(this,0.0,1.0,numberOfSpectra);
 
   PARALLEL_FOR2(inputWS,outputWS)
-  for (int64_t i = 0; i < numberOfSpectra; ++i)
+  for (int64_t i = 0; i < int64_t(numberOfSpectra); ++i)
   {
     PARALLEL_START_INTERUPT_REGION
     // Get the spectrum number for this histogram
